chore: apply padding to parent routes component
